unsigned hustle

+++ 0x00-hello_world/6-size.c
@@ -14,9 +14,9 @@ long long int c;
 char d;
 float f;
 
-printf("size of a char: %1u byte(s)\n", (unsigned long)sizeof(d));
-printf("size of an int: %1u byte(s)\n", (unsigned long)sizeof(a));
-printf("size of a long int: %1u byte(s)\n", (unsigned long)sizeof(b));
+printf("size of a char: %1u byte(s)\n", (unsigned)sizeof(d));
+printf("size of an int: %1u byte(s)\n", (unsigned)sizeof(a));
+printf("size of a long int: %1u byte(s)\n", (unsigned)sizeof(b));
 printf("size of a long long int: %1u byte(s)\n", (unsigned long)sizeof(c));
 printf("size of a float: %1u byte(s)\n", (unsigned long)sizeof(f));
 fprintf(stderr, "");
